clearing ref of post after delete from user done

+++ controllers/feed.js
@@ -129,6 +129,11 @@ exports.updatePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             };
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not Authorized!');
+                error.statusCode = 403;
+                throw error;
+            }
             if (imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl);
             }
@@ -167,12 +172,23 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             };
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not Authorized!');
+                error.statusCode = 403;
+                throw error;
+            }
             //checked logged in user
             clearImage(post.imageUrl)
             return Post.findByIdAndRemove(postId);
         })
         .then(result => {
-            console.log(result);
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            user.posts.pull(postId);
+            return user.save()
+        })
+        .then(result => {
             res.status(200).json({
                 message: 'Post Deleted Successfully.'
             })
